Another fix for previous commit
Forgot to change json type from int to float. Fixed now.

+++ gdx/src/com/badlogic/gdx/scenes/scene2d/ui/Skin.java
@@ -532,7 +532,7 @@ public class Skin implements Disposable {
 		json.setSerializer(BitmapFont.class, new ReadOnlySerializer<BitmapFont>() {
 			public BitmapFont read (Json json, JsonValue jsonData, Class type) {
 				String path = json.readValue("file", String.class, jsonData);
-				float scaledSize = json.readValue("scaledSize", int.class, -1, jsonData);
+				float scaledSize = json.readValue("scaledSize", float.class, -1f, jsonData);
 				Boolean flip = json.readValue("flip", Boolean.class, false, jsonData);
 				Boolean markupEnabled = json.readValue("markupEnabled", Boolean.class, false, jsonData);
 				Boolean integerPositions = json.readValue("integerPositions", Boolean.class, true, jsonData);
